pyproject.toml
[project]
name = "qa-playwright-python-framework"
version = "0.1.0"
description = ""
authors = [
    {name = "uadmon", email = "qa-automation@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pytest-playwright (>=0.7.2,<0.8.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"